requirements.txt
asgiref==3.2.10
astroid==2.4.2
certifi==2020.6.20
chardet==3.0.4
Django==3.1.1
django-appconf==1.0.4
django-filter==2.3.0
djangorestframework==3.11.1
flake8==3.8.3
idna==2.10
isort==5.4.2
lazy-object-proxy==1.4.3
Markdown==3.2.2
mccabe==0.6.1
Pillow==7.2.0
pinax-likes==4.0.0
pyflakes==2.2.0
pyroute2==0.5.13
pytz==2020.1
requests==2.24.0
six==1.15.0
sqlparse==0.3.1
toml==0.10.1
urllib3==1.25.10
wrapt==1.12.1
